jwt_simple

Module prelude

Re-exports

pub use ct_codecs::Decoder as _;

pub use ct_codecs::Encoder as _;

pub use crate::jwe_token::DecryptionOptions;

pub use crate::jwe_token::EncryptionOptions;

pub use crate::jwe_token::JWEToken;

pub use crate::jwe_token::JWETokenMetadata;

pub use hashset_from_strings::HashSetFromStringsT as _;

pub use coarsetime;

pub use crate::algorithms::*;

pub use crate::claims::*;

pub use crate::common::*;

pub use crate::token::*;

Structs

Base64

Standard Base64 encoder and decoder with padding.

Base64NoPadding

Standard Base64 encoder and decoder without padding.

Base64UrlSafe

URL-safe Base64 encoder and decoder with padding.

Base64UrlSafeNoPadding

URL-safe Base64 encoder and decoder without padding.

Clock

System time

Duration

A duration type to represent an approximate span of time

HashSet

A hash set implemented as a HashMap where the value is ().

Traits

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Type Aliases

UnixTimeStamp

Alias for Duration.

Derive Macros

Deserialize

Serialize