jwt_simple::prelude

Struct Duration

pub struct Duration(/* private fields */);

A duration type to represent an approximate span of time

Implementations

impl Duration

pub const fn new(sec: u64, nanos: u32) -> Duration

Creates a new Duration from the specified number of seconds and additional nanosecond precision

pub const fn from_days(days: u64) -> Duration

Creates a new Duration from the specified number of days

pub const fn from_hours(hours: u64) -> Duration

Creates a new Duration from the specified number of hours

pub const fn from_mins(mins: u64) -> Duration

Creates a new Duration from the specified number of minutes

pub const fn from_secs(secs: u64) -> Duration

Creates a new Duration from the specified number of seconds

pub const fn from_millis(millis: u64) -> Duration

Creates a new Duration from the specified number of milliseconds

pub const fn as_days(&self) -> u64

Returns the number of days represented by this duration

pub const fn as_hours(&self) -> u64

Returns the number of hours represented by this duration

pub const fn as_mins(&self) -> u64

Returns the number of minutes represented by this duration

pub const fn as_secs(&self) -> u64

Returns the number of whole seconds represented by this duration

pub const fn as_millis(&self) -> u64

Returns the number of whole milliseconds represented by this duration

pub const fn as_micros(&self) -> u64

Returns the number of whole microseconds represented by this duration

pub const fn as_nanos(&self) -> u64

Returns the number of whole nanoseconds represented by this duration

pub const fn subsec_nanos(&self) -> u32

Returns the nanosecond precision represented by this duration

pub const fn as_ticks(&self) -> u64

Return this duration as a number of “ticks”.

Note that length of a ‘tick’ is not guaranteed to represent the same amount of time across different platforms, or from one version of coarsetime to another.

pub const fn from_ticks(ticks: u64) -> Duration

Creates a new Duration from the specified number of “ticks”.

Note that length of a ‘tick’ is not guaranteed to represent the same amount of time across different platforms, or from one version of coarsetime to another.

pub fn as_f64(&self) -> f64

Returns the duration as a floating point number, representing the number of seconds

pub const fn abs_diff(&self, other: Duration) -> Duration

Returns the absolute difference between two Durations

pub const fn saturating_add(self, rhs: Duration) -> Duration

Add two durations, saturating on overflow

pub fn checked_add(self, rhs: Duration) -> Option<Duration>

Add two durations, returning None on overflow

pub const fn saturating_sub(self, rhs: Duration) -> Duration

Subtract two durations, saturating on underflow/overflow

pub fn checked_sub(self, rhs: Duration) -> Option<Duration>

Subtract two durations, returning None on underflow/overflow

pub const fn saturating_mul(self, rhs: u32) -> Duration

Multiply a duration by a scalar, saturating on overflow

pub fn checked_mul(self, rhs: u32) -> Option<Duration>

Multiply a duration by a scalar, returning None on overflow

pub fn checked_div(self, rhs: u32) -> Option<Duration>

Divide a duration by a scalar, returning None for division by zero